Zoom the x-axis range for a normalizedMatrix coverage data

Zoom the x-axis range for a normalizedMatrix coverage data

Usage

zoom_nmat(nmat, upstream_length = 500, downstream_length = 500, ...)

Arguments

nmat

normalizedMatrix object, where the length extended from the target region is stored in attr(nmat, "extend") as a two-element integer vector representing upstream, and downstream length. Each column indicated in attr(nmat, "upstream_index") is expected to represent equal-sized bins spanning that range. Columns are retained if the farthest distance of the column is less than upstream_length.

upstream_length, downstream_length

numeric coordinate maximum range from the target center region. When either value is NULL no threshold is applied, which is equivalent to Inf. The values are forced positive abs(upstream_length) as these are absolute magnitude length from the target region.

...

additional arguments are ignored.

Details

This function is typically called by zoom_nmatlist() but can be called on an individual normalizedMatrix object.

This function filters the matrix columns by distance, and updates important associated attributes:

attr(nmat, "upstream_index") - the column index positions upstream the target region

attr(nmat, "downstream_index") - the column index positions downstream the target region

attr(nmat, "target_index") - the column index positions representing the target region

attr(nmat, "extend") - the genomic distance upstream and downstream the target region

See also